refactor: union visitors
