+++ crates/swc_css_compat/src/utils.rs
@@ -1,0 +1,5 @@
+use swc_css_ast::{ComponentValue, Rule};
+
+pub(crate) fn rule_to_component_value(rule: Rule) -> ComponentValue {
+    rule.into()
+}
